move range to the utils - prevent shadowing built-ins
